consolidate all operations into db_ops
