+++ lib/db.js
@@ -3,13 +3,11 @@
 const EventEmitter = require('events').EventEmitter;
 const inherits = require('util').inherits;
 const getSingleProperty = require('./utils').getSingleProperty;
-const shallowClone = require('./utils').shallowClone;
 const CommandCursor = require('./command_cursor');
 const handleCallback = require('./utils').handleCallback;
 const filterOptions = require('./utils').filterOptions;
 const toError = require('./utils').toError;
 const ReadPreference = require('mongodb-core').ReadPreference;
-const f = require('util').format;
 const Admin = require('./admin');
 const MongoError = require('mongodb-core').MongoError;
 const ObjectID = require('mongodb-core').ObjectID;
@@ -18,26 +16,27 @@ const Collection = require('./collection');
 const mergeOptionsAndWriteConcern = require('./utils').mergeOptionsAndWriteConcern;
 const executeOperation = require('./utils').executeOperation;
 const applyWriteConcern = require('./utils').applyWriteConcern;
+const convertReadPreference = require('./utils').convertReadPreference;
 
 // Operations
-const addUser = require('./operations/db_user_ops').addUser;
-const collections = require('./operations/db_collections');
-const createCollection = require('./operations/db_collection_ops').createCollection;
-const createIndex = require('./operations/db_index_ops').createIndex;
-const dropCollection = require('./operations/db_collection_ops').dropCollection;
-const dropDatabase = require('./operations/db_drop_database');
-const ensureIndex = require('./operations/db_index_ops').ensureIndex;
-const evaluate = require('./operations/db_evaluate');
-const executeCommand = require('./operations/db_execute_commands').executeCommand;
-const executeDbAdminCommand = require('./operations/db_execute_commands').executeDbAdminCommand;
-const indexInformation = require('./operations/db_index_ops').indexInformation;
-const removeUser = require('./operations/db_user_ops').removeUser;
-
-// Helpers
-const convertReadPreference = require('./operations/db_helpers').convertReadPreference;
-const profilingInfo = require('./operations/db_helpers').profilingInfo;
-const profilingLevel = require('./operations/db_helpers').profilingLevel;
-const setProfilingLevel = require('./operations/db_helpers').setProfilingLevel;
+const addUser = require('./operations/db_ops').addUser;
+const collections = require('./operations/db_ops').collections;
+const createCollection = require('./operations/db_ops').createCollection;
+const createIndex = require('./operations/db_ops').createIndex;
+const createListener = require('./operations/db_ops').createListener;
+const dropCollection = require('./operations/db_ops').dropCollection;
+const dropDatabase = require('./operations/db_ops').dropDatabase;
+const ensureIndex = require('./operations/db_ops').ensureIndex;
+const evaluate = require('./operations/db_ops').evaluate;
+const executeCommand = require('./operations/db_ops').executeCommand;
+const executeDbAdminCommand = require('./operations/db_ops').executeDbAdminCommand;
+const indexInformation = require('./operations/db_ops').indexInformation;
+const listCollectionsTransforms = require('./operations/db_ops').listCollectionsTransforms;
+const profilingInfo = require('./operations/db_ops').profilingInfo;
+const profilingLevel = require('./operations/db_ops').profilingLevel;
+const removeUser = require('./operations/db_ops').removeUser;
+const setProfilingLevel = require('./operations/db_ops').setProfilingLevel;
+const validateDatabaseName = require('./operations/db_ops').validateDatabaseName;
 
 /**
  * @fileOverview The **Db** class is a class that represents a MongoDB Database.
@@ -57,7 +56,7 @@ const setProfilingLevel = require('./operations/db_helpers').setProfilingLevel;
  */
 
 // Allowed parameters
-var legalOptionNames = [
+const legalOptionNames = [
   'w',
   'wtimeout',
   'fsync',
@@ -127,14 +126,13 @@ var legalOptionNames = [
  * @fires Db#fullsetup
  * @return {Db} a Db instance.
  */
-var Db = function(databaseName, topology, options) {
+const Db = function(databaseName, topology, options) {
   options = options || {};
   if (!(this instanceof Db)) return new Db(databaseName, topology, options);
   EventEmitter.call(this);
-  var self = this;
 
   // Get the promiseLibrary
-  var promiseLibrary = options.promiseLibrary || Promise;
+  const promiseLibrary = options.promiseLibrary || Promise;
 
   // Filter the options
   options = filterOptions(options, legalOptionNames);
@@ -142,7 +140,7 @@ var Db = function(databaseName, topology, options) {
   // Ensure we put the promiseLib in the options
   options.promiseLibrary = promiseLibrary;
 
-  // var self = this;  // Internal state of the db object
+  // Internal state of the db object
   this.s = {
     // Database name
     databaseName: databaseName,
@@ -177,26 +175,26 @@ var Db = function(databaseName, topology, options) {
   };
 
   // Ensure we have a valid db name
-  validateDatabaseName(self.s.databaseName);
+  validateDatabaseName(this.s.databaseName);
 
   // Add a read Only property
-  getSingleProperty(this, 'serverConfig', self.s.topology);
-  getSingleProperty(this, 'bufferMaxEntries', self.s.bufferMaxEntries);
-  getSingleProperty(this, 'databaseName', self.s.databaseName);
+  getSingleProperty(this, 'serverConfig', this.s.topology);
+  getSingleProperty(this, 'bufferMaxEntries', this.s.bufferMaxEntries);
+  getSingleProperty(this, 'databaseName', this.s.databaseName);
 
   // This is a child db, do not register any listeners
   if (options.parentDb) return;
   if (this.s.noListener) return;
 
   // Add listeners
-  topology.on('error', createListener(self, 'error', self));
-  topology.on('timeout', createListener(self, 'timeout', self));
-  topology.on('close', createListener(self, 'close', self));
-  topology.on('parseError', createListener(self, 'parseError', self));
-  topology.once('open', createListener(self, 'open', self));
-  topology.once('fullsetup', createListener(self, 'fullsetup', self));
-  topology.once('all', createListener(self, 'all', self));
-  topology.on('reconnect', createListener(self, 'reconnect', self));
+  topology.on('error', createListener(this, 'error', this));
+  topology.on('timeout', createListener(this, 'timeout', this));
+  topology.on('close', createListener(this, 'close', this));
+  topology.on('parseError', createListener(this, 'parseError', this));
+  topology.once('open', createListener(this, 'open', this));
+  topology.once('fullsetup', createListener(this, 'fullsetup', this));
+  topology.once('all', createListener(this, 'all', this));
+  topology.on('reconnect', createListener(this, 'reconnect', this));
 };
 
 inherits(Db, EventEmitter);
@@ -236,7 +234,7 @@ Object.defineProperty(Db.prototype, 'slaveOk', {
 Object.defineProperty(Db.prototype, 'writeConcern', {
   enumerable: true,
   get: function() {
-    var ops = {};
+    const ops = {};
     if (this.s.options.w != null) ops.w = this.s.options.w;
     if (this.s.options.j != null) ops.j = this.s.options.j;
     if (this.s.options.fsync != null) ops.fsync = this.s.options.fsync;
@@ -256,10 +254,8 @@ Object.defineProperty(Db.prototype, 'writeConcern', {
  * @return {Promise} returns Promise if no callback passed
  */
 Db.prototype.command = function(command, options, callback) {
-  // Change the callback
   if (typeof options === 'function') (callback = options), (options = {});
-  // Clone the options
-  options = shallowClone(options);
+  options = Object.assign({}, options);
 
   return executeOperation(this.s.topology, executeCommand, [this, command, options, callback]);
 };
@@ -280,7 +276,7 @@ Db.prototype.admin = function() {
  * @param {Collection} collection The collection instance.
  */
 
-var collectionKeys = [
+const collectionKeys = [
   'pkFactory',
   'readPreference',
   'serializeFunctions',
@@ -294,7 +290,7 @@ var collectionKeys = [
 
 /**
  * Fetch a specific collection (containing the actual collection information). If the application does not use strict mode you
- * can use it without a callback in the following way: `var collection = db.collection('mycollection');`
+ * can use it without a callback in the following way: `const collection = db.collection('mycollection');`
  *
  * @method
  * @param {string} name the collection name we wish to access.
@@ -315,10 +311,10 @@ var collectionKeys = [
  * @return {Collection} return the new Collection instance if not in strict mode
  */
 Db.prototype.collection = function(name, options, callback) {
-  var self = this;
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
-  options = shallowClone(options);
+  Object.assign({}, options);
+
   // Set the promise library
   options.promiseLibrary = this.s.promiseLibrary;
 
@@ -336,7 +332,7 @@ Db.prototype.collection = function(name, options, callback) {
   // Execute
   if (options == null || !options.strict) {
     try {
-      var collection = new Collection(
+      const collection = new Collection(
         this,
         this.s.topology,
         this.s.databaseName,
@@ -354,23 +350,23 @@ Db.prototype.collection = function(name, options, callback) {
 
   // Strict mode
   if (typeof callback !== 'function') {
-    throw toError(f('A callback is required in strict mode. While getting collection %s.', name));
+    throw toError(`A callback is required in strict mode. While getting collection ${name}`);
   }
 
   // Did the user destroy the topology
-  if (self.serverConfig && self.serverConfig.isDestroyed()) {
+  if (this.serverConfig && this.serverConfig.isDestroyed()) {
     return callback(new MongoError('topology was destroyed'));
   }
 
   const listCollectionOptions = Object.assign({}, options, { nameOnly: true });
 
   // Strict mode
-  this.listCollections({ name: name }, listCollectionOptions).toArray(function(err, collections) {
+  this.listCollections({ name: name }, listCollectionOptions).toArray((err, collections) => {
     if (err != null) return handleCallback(callback, err, null);
     if (collections.length === 0)
       return handleCallback(
         callback,
-        toError(f('Collection %s does not exist. Currently in strict mode.', name)),
+        toError(`Collection ${name} does not exist. Currently in strict mode.`),
         null
       );
 
@@ -378,7 +374,7 @@ Db.prototype.collection = function(name, options, callback) {
       return handleCallback(
         callback,
         null,
-        new Collection(self, self.s.topology, self.s.databaseName, name, self.s.pkFactory, options)
+        new Collection(this, this.s.topology, this.s.databaseName, name, this.s.pkFactory, options)
       );
     } catch (err) {
       return handleCallback(callback, err, null);
@@ -444,7 +440,7 @@ Db.prototype.stats = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
   // Build command object
-  var commandObject = { dbStats: true };
+  const commandObject = { dbStats: true };
   // Check if we have the scale value
   if (options['scale'] != null) commandObject['scale'] = options['scale'];
 
@@ -455,23 +451,6 @@ Db.prototype.stats = function(options, callback) {
 
   // Execute the command
   return this.command(commandObject, options, callback);
-};
-
-// Transformation methods for cursor results
-var listCollectionsTranforms = function(databaseName) {
-  var matching = f('%s.', databaseName);
-
-  return {
-    doc: function(doc) {
-      var index = doc.name.indexOf(matching);
-      // Remove database name if available
-      if (doc.name && index === 0) {
-        doc.name = doc.name.substr(index + matching.length);
-      }
-
-      return doc;
-    }
-  };
 };
 
 /**
@@ -491,7 +470,7 @@ Db.prototype.listCollections = function(filter, options) {
   options = options || {};
 
   // Shallow clone the object
-  options = shallowClone(options);
+  options = Object.assign({}, options);
   // Set the promise library
   options.promiseLibrary = this.s.promiseLibrary;
 
@@ -502,18 +481,18 @@ Db.prototype.listCollections = function(filter, options) {
     options.readPreference = this.s.readPreference || ReadPreference.primary;
   }
 
+  // Cursor options
+  let cursor = options.batchSize ? { batchSize: options.batchSize } : {};
+
   // We have a list collections command
   if (this.serverConfig.capabilities().hasListCollectionsCommand) {
-    // Cursor options
-    var cursor = options.batchSize ? { batchSize: options.batchSize } : {};
-
     const nameOnly = typeof options.nameOnly === 'boolean' ? options.nameOnly : false;
     // Build the command
-    var command = { listCollections: true, filter, cursor, nameOnly };
+    const command = { listCollections: true, filter, cursor, nameOnly };
     // Set the AggregationCursor constructor
     options.cursorFactory = CommandCursor;
     // Create the cursor
-    cursor = this.s.topology.cursor(f('%s.$cmd', this.s.databaseName), command, options);
+    cursor = this.s.topology.cursor(`${this.s.databaseName}.$cmd`, command, options);
     // Do we have a readPreference, apply it
     if (options.readPreference) {
       cursor.setReadPreference(options.readPreference);
@@ -529,14 +508,14 @@ Db.prototype.listCollections = function(filter, options) {
       typeof filter.name === 'string' &&
       !new RegExp('^' + this.databaseName + '\\.').test(filter.name)
     ) {
-      filter = shallowClone(filter);
-      filter.name = f('%s.%s', this.s.databaseName, filter.name);
+      filter = Object.assign({}, filter);
+      filter.name = `${this.s.databaseName}.${filter.name}`;
     }
   }
 
   // No filter, filter by current database
   if (filter == null) {
-    filter.name = f('/%s/', this.s.databaseName);
+    filter.name = `/${this.s.databaseName}/`;
   }
 
   // Rewrite the filter to use $and to filter out indexes
@@ -547,7 +526,7 @@ Db.prototype.listCollections = function(filter, options) {
   }
 
   // Return options
-  var _options = { transforms: listCollectionsTranforms(this.s.databaseName) };
+  const _options = { transforms: listCollectionsTransforms(this.s.databaseName) };
   // Get the cursor
   cursor = this.collection(Db.SYSTEM_NAMESPACE_COLLECTION).find(filter, _options);
   // Do we have a readPreference, apply it
@@ -572,7 +551,7 @@ Db.prototype.listCollections = function(filter, options) {
  * @return {Promise} returns Promise if no callback passed
  */
 Db.prototype.eval = function(code, parameters, options, callback) {
-  var args = Array.prototype.slice.call(arguments, 1);
+  const args = Array.prototype.slice.call(arguments, 1);
   callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
   parameters = args.length ? args.shift() : parameters;
   options = args.length ? args.shift() || {} : {};
@@ -621,7 +600,7 @@ Db.prototype.dropCollection = function(name, options, callback) {
   options = options || {};
 
   // Command to execute
-  var cmd = { drop: name };
+  const cmd = { drop: name };
 
   // Decorate with write concern
   applyWriteConcern(cmd, { db: this }, options);
@@ -646,7 +625,7 @@ Db.prototype.dropDatabase = function(options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
   options = options || {};
   // Drop database command
-  var cmd = { dropDatabase: 1 };
+  const cmd = { dropDatabase: 1 };
 
   // Decorate with write concern
   applyWriteConcern(cmd, { db: this }, options);
@@ -707,7 +686,7 @@ Db.prototype.executeDbAdminCommand = function(selector, options, callback) {
 };
 
 /**
- * Creates an index on the db and collection collection.
+ * Creates an index on the db and collection.
  * @method
  * @param {string} name Name of the collection to create the index on.
  * @param {(string|object)} fieldOrSpec Defines the index.
@@ -731,7 +710,7 @@ Db.prototype.executeDbAdminCommand = function(selector, options, callback) {
  */
 Db.prototype.createIndex = function(name, fieldOrSpec, options, callback) {
   if (typeof options === 'function') (callback = options), (options = {});
-  options = options ? shallowClone(options) : {};
+  options = options ? Object.assign({}, options) : {};
 
   return executeOperation(this.s.topology, createIndex, [
     this,
@@ -799,16 +778,9 @@ Db.prototype.addChild = function(db) {
  * @return {Promise} returns Promise if no callback passed
  */
 Db.prototype.addUser = function(username, password, options, callback) {
-  // Unpack the parameters
-  // Change the callback
   if (typeof options === 'function') (callback = options), (options = {});
-  // Clone the options
-  options = shallowClone(options);
-  /*
-  var args = Array.prototype.slice.call(arguments, 2);
-  callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
-  options = args.length ? args.shift() || {} : {};
-  */
+  options = options || {};
+
   return executeOperation(this.s.topology, addUser, [this, username, password, options, callback]);
 };
 
@@ -825,14 +797,9 @@ Db.prototype.addUser = function(username, password, options, callback) {
  * @return {Promise} returns Promise if no callback passed
  */
 Db.prototype.removeUser = function(username, options, callback) {
-  /*var args = Array.prototype.slice.call(arguments, 1);
-  callback = typeof args[args.length - 1] === 'function' ? args.pop() : undefined;
-  options = args.length ? args.shift() || {} : {};
-*/
-  // Change the callback
   if (typeof options === 'function') (callback = options), (options = {});
-  // Clone the options
-  options = shallowClone(options);
+  options = options || {};
+
   return executeOperation(this.s.topology, removeUser, [this, username, options, callback]);
 };
 
@@ -899,39 +866,6 @@ Db.prototype.indexInformation = function(name, options, callback) {
   options = options || {};
 
   return executeOperation(this.s.topology, indexInformation, [this, name, options, callback]);
-};
-
-// Validate the database name
-var validateDatabaseName = function(databaseName) {
-  if (typeof databaseName !== 'string')
-    throw MongoError.create({ message: 'database name must be a string', driver: true });
-  if (databaseName.length === 0)
-    throw MongoError.create({ message: 'database name cannot be the empty string', driver: true });
-  if (databaseName === '$external') return;
-
-  var invalidChars = [' ', '.', '$', '/', '\\'];
-  for (var i = 0; i < invalidChars.length; i++) {
-    if (databaseName.indexOf(invalidChars[i]) !== -1)
-      throw MongoError.create({
-        message: "database names cannot contain the character '" + invalidChars[i] + "'",
-        driver: true
-      });
-  }
-};
-
-// Add listeners to topology
-var createListener = function(self, e, object) {
-  var listener = function(err) {
-    if (object.listeners(e).length > 0) {
-      object.emit(e, err, self);
-
-      // Emit on all associated db's if available
-      for (var i = 0; i < self.s.children.length; i++) {
-        self.s.children[i].emit(e, err, self.s.children[i]);
-      }
-    }
-  };
-  return listener;
 };
 
 /**
