+++ lib/utils.js
@@ -496,6 +496,34 @@ function applyWriteConcern(target, sources, options) {
 }
 
 /**
+ * Ensures provided read preference is properly converted into an object
+ * @param {(ReadPreference|string|object)} readPreference the user provided read preference
+ * @return {ReadPreference}
+ */
+function convertReadPreference(readPreference) {
+  if (readPreference) {
+    if (typeof readPreference === 'string') {
+      return new ReadPreference(readPreference);
+    } else if (
+      readPreference &&
+      !(readPreference instanceof ReadPreference) &&
+      typeof readPreference === 'object'
+    ) {
+      const mode = readPreference.mode || readPreference.preference;
+      if (mode && typeof mode === 'string') {
+        return new ReadPreference(mode, readPreference.tags, {
+          maxStalenessSeconds: readPreference.maxStalenessSeconds
+        });
+      }
+    } else if (!(readPreference instanceof ReadPreference)) {
+      throw new TypeError('Invalid read preference: ' + readPreference);
+    }
+  }
+
+  return readPreference;
+}
+
+/**
  * Checks if a given value is a Promise
  *
  * @param {*} maybePromise
@@ -505,23 +533,26 @@ function isPromiseLike(maybePromise) {
   return maybePromise && typeof maybePromise.then === 'function';
 }
 
-exports.filterOptions = filterOptions;
-exports.mergeOptions = mergeOptions;
-exports.translateOptions = translateOptions;
-exports.shallowClone = shallowClone;
-exports.getSingleProperty = getSingleProperty;
-exports.checkCollectionName = checkCollectionName;
-exports.toError = toError;
-exports.formattedOrderClause = formattedOrderClause;
-exports.parseIndexOptions = parseIndexOptions;
-exports.normalizeHintField = normalizeHintField;
-exports.handleCallback = handleCallback;
-exports.decorateCommand = decorateCommand;
-exports.isObject = isObject;
-exports.debugOptions = debugOptions;
-exports.MAX_JS_INT = 0x20000000000000;
-exports.mergeOptionsAndWriteConcern = mergeOptionsAndWriteConcern;
-exports.translateReadPreference = translateReadPreference;
-exports.executeOperation = executeOperation;
-exports.applyWriteConcern = applyWriteConcern;
-exports.isPromiseLike = isPromiseLike;
+module.exports = {
+  filterOptions,
+  mergeOptions,
+  translateOptions,
+  shallowClone,
+  getSingleProperty,
+  checkCollectionName,
+  toError,
+  formattedOrderClause,
+  parseIndexOptions,
+  normalizeHintField,
+  handleCallback,
+  decorateCommand,
+  isObject,
+  debugOptions,
+  MAX_JS_INT: 0x20000000000000,
+  mergeOptionsAndWriteConcern,
+  translateReadPreference,
+  executeOperation,
+  applyWriteConcern,
+  convertReadPreference,
+  isPromiseLike
+};
